requirements.txt
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.2.2